XWIKI-15787: Add a "Clear all" button on the recycle bin
  * Delegate to the refactoring job the role of getting all documents
from the recycle bin in batches and deleting them

+++ xwiki-platform-core/xwiki-platform-oldcore/src/main/java/com/xpn/xwiki/store/XWikiRecycleBinStoreInterface.java
@@ -163,9 +163,9 @@ public interface XWikiRecycleBinStoreInterface
      * @throws XWikiException - if error in loading.
      * @since 10.10RC1
      */
-    default int getNumberOfDeletedDocuments(XWikiContext context) throws XWikiException
+    default Long getNumberOfDeletedDocuments(XWikiContext context) throws XWikiException
     {
-        return -1;
+        return -1L;
     }
 
     /**
